requirements.txt
textual
mutagen
python-vlc
